decouple message rendering to its own component, can be further split though
