+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/seo-assistant/seo-assistant-wizard.tsx
@@ -1,21 +1,13 @@
-import {
-	useState,
-	useCallback,
-	useEffect,
-	useRef,
-	useMemo,
-	// createInterpolateElement,
-} from '@wordpress/element';
+import { useState, useCallback, useEffect, useRef, useMemo } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
-import clsx from 'clsx';
 import debugFactory from 'debug';
 import './style.scss';
-import bigSkyIcon from './big-sky-icon.svg';
 import { useCompletionStep } from './use-completion-step';
 import { useKeywordsStep } from './use-keywords-step';
 import { useMetaDescriptionStep } from './use-meta-description-step';
 import { useTitleStep } from './use-title-step';
 import WizardInput from './wizard-input';
+import WizardMessages from './wizard-messages';
 import type { SeoAssistantProps, Step, Message } from './types';
 
 const debug = debugFactory( 'jetpack-ai:seo-assistant-wizard' );
@@ -204,72 +196,6 @@ export default function SeoAssistantWizard( { isOpen, close, onStep }: SeoAssist
 			.forEach( step => step.setCompleted( false ) );
 	}, [ close, steps ] );
 
-	const renderMessageText = message => {
-		if ( message.type === 'past-options' ) {
-			return (
-				<div className="seo-assistant-wizard__options">
-					{ message.options.map( option => (
-						<div
-							key={ option.id }
-							className={ clsx( 'seo-assistant-wizard__option', {
-								'is-selected': option.selected,
-							} ) }
-						>
-							{ option.content }
-						</div>
-					) ) }
-				</div>
-			);
-		}
-
-		return <div className="seo-assistant-wizard__message-text">{ message.content }</div>;
-	};
-
-	const renderMessages = () => {
-		return messages.map( message => (
-			<div
-				key={ message.id }
-				className={ clsx( 'seo-assistant-wizard__message', {
-					'is-user': message.isUser,
-				} ) }
-			>
-				<div className="seo-assistant-wizard__message-icon">
-					{ message.showIcon && (
-						<img src={ bigSkyIcon } alt={ __( 'SEO Assistant avatar', 'jetpack' ) } />
-					) }
-				</div>
-				{ renderMessageText( message ) }
-			</div>
-		) );
-	};
-
-	const renderOptions = () => {
-		if ( currentStepData.type !== 'options' || ! currentStepData.options.length ) {
-			return null;
-		}
-
-		return (
-			<div className="seo-assistant-wizard__message">
-				<div className="seo-assistant-wizard__message-icon"></div>
-				<div className="seo-assistant-wizard__message-text">
-					<div className="seo-assistant-wizard__options">
-						{ currentStepData.options.map( option => (
-							<button
-								key={ option.id }
-								className={ clsx( 'seo-assistant-wizard__option', {
-									'is-selected': option.selected,
-								} ) }
-								onClick={ () => currentStepData.onSelect( option ) }
-							>
-								{ option.content }
-							</button>
-						) ) }
-					</div>
-				</div>
-			</div>
-		);
-	};
-
 	return (
 		isOpen && (
 			<div className="seo-assistant-wizard">
@@ -284,11 +210,7 @@ export default function SeoAssistantWizard( { isOpen, close, onStep }: SeoAssist
 				</div>
 
 				<div className="seo-assistant-wizard__content">
-					<div className="seo-assistant-wizard__messages">
-						{ renderMessages() }
-						{ renderOptions() }
-						<div ref={ messagesEndRef } />
-					</div>
+					<WizardMessages currentStepData={ currentStepData } messages={ messages } />
 
 					<WizardInput
 						currentStepData={ currentStepData }
